updated add_ctf command, added a utilities file and set_creds functionality.
This first proper commit is an attempt of trying to create much better documentation of what I have done
with this side-project of mine. Instead of creating text channels for CTFs, I have decided to instead create
a seperate category and create an __init__ channel wherein the main team communication will take place. Most
of the features that I will include in future updates can be found on the README.